# backend/openapi.json
{
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "network_footprint",
          "technology_stack",
          "human_layer",
          "leaked_data"
        ],
        "title": "Category",
        "type": "string"
      },
      "CollectorRunRead": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "cache_hit": {
            "title": "Cache Hit",
            "type": "boolean"
          },
          "collector": {
            "title": "Collector",
            "type": "string"
          },
          "finding_count": {
            "title": "Finding Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "safe_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Code"
          },
          "safe_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/CollectorStatus"
          }
        },
        "required": [
          "collector",
          "status",
          "attempt_count",
          "cache_hit",
          "finding_count",
          "safe_error_code",
          "safe_error_message",
          "started_at",
          "finished_at"
        ],
        "title": "CollectorRunRead",
        "type": "object"
      },
      "CollectorStatus": {
        "enum": [
          "queued",
          "running",
          "done",
          "failed",
          "skipped_no_key",
          "not_applicable",
          "interrupted"
        ],
        "title": "CollectorStatus",
        "type": "string"
      },
      "Confidence": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Confidence",
        "type": "string"
      },
...
      "FindingRead": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "collector": {
            "title": "Collector",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "normalized_value": {
            "additionalProperties": true,
            "title": "Normalized Value",
            "type": "object"
          },
          "provider_observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Observed At"
          },
          "raw_evidence": {
            "additionalProperties": true,
            "title": "Raw Evidence",
            "type": "object"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "collector",
          "category",
          "kind",
          "title",
          "summary",
          "normalized_value",
          "raw_evidence",
          "source_url",
          "provider_observed_at",
          "retrieved_at",
          "confidence",
          "fingerprint"
        ],
        "title": "FindingRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobCreateRequest": {
        "properties": {
          "attestation_confirmed": {
            "title": "Attestation Confirmed",
            "type": "boolean"
          },
          "scope_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Note"
          },
          "selected_sources": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Selected Sources",
            "type": "array"
          },
          "target": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "selected_sources",
          "attestation_confirmed"
        ],
        "title": "JobCreateRequest",
        "type": "object"
      },
      "JobDetail": {
        "properties": {
          "attestation_text": {
            "title": "Attestation Text",
            "type": "string"
          },
          "attestation_time": {
            "format": "date-time",
            "title": "Attestation Time",
            "type": "string"
          },
          "attestation_version": {
            "title": "Attestation Version",
            "type": "string"
          },
          "collector_runs": {
            "items": {
              "$ref": "#/components/schemas/CollectorRunRead"
            },
            "title": "Collector Runs",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "scope_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Note"
          },
          "selected_sources": {
            "items": {
              "type": "string"
            },
            "title": "Selected Sources",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "target_input": {
            "title": "Target Input",
            "type": "string"
          },
          "target_normalized": {
            "title": "Target Normalized",
            "type": "string"
          },
          "target_type": {
            "$ref": "#/components/schemas/TargetType"
          },
          "warning_count": {
            "default": 0,
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "target_input",
          "target_normalized",
          "target_type",
          "status",
          "selected_sources",
          "scope_note",
          "created_at",
          "started_at",
          "finished_at",
          "attestation_text",
          "attestation_version",
          "attestation_time",
          "collector_runs"
        ],
        "title": "JobDetail",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "completed_with_warnings",
          "failed",
          "canceled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "scope_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Note"
          },
          "selected_sources": {
            "items": {
              "type": "string"
            },
            "title": "Selected Sources",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "target_input": {
            "title": "Target Input",
            "type": "string"
          },
          "target_normalized": {
            "title": "Target Normalized",
            "type": "string"
          },
          "target_type": {
            "$ref": "#/components/schemas/TargetType"
          },
          "warning_count": {
            "default": 0,
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "target_input",
          "target_normalized",
          "target_type",
          "status",
          "selected_sources",
          "scope_note",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobSummary",
        "type": "object"
      },
...
      "SubdomainRowRead": {
        "properties": {
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "in_scope": {
            "title": "In Scope",
            "type": "boolean"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "subdomain": {
            "title": "Subdomain",
            "type": "string"
          },
          "wildcard": {
            "title": "Wildcard",
            "type": "boolean"
          }
        },
        "required": [
          "subdomain",
          "source_count",
          "sources",
          "first_seen_at",
          "last_seen_at",
          "wildcard",
          "in_scope"
        ],
        "title": "SubdomainRowRead",
        "type": "object"
      },
      "TargetType": {
        "enum": [
          "domain",
          "ip",
          "cidr",
          "organization"
        ],
        "title": "TargetType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ReconLedger",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/jobs": {
      "get": {
        "description": "UX-09 history: target, type, status, time, selected sources, and a\nwarning count (failed collector runs) - the last one computed here via\nan aggregate join rather than loading every job's full collector_runs\nlist, which is what makes JobDetail heavier than JobSummary.",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_endpoint_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Job Api Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/events": {
      "get": {
        "description": "SSE is advisory (7.3): this just streams job_events rows as they\nappear. GET /api/jobs/{id} remains authoritative, and Last-Event-ID\nsupport lets a client resume a dropped connection without missing events.",
        "operationId": "job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/findings": {
      "get": {
        "description": "UX-08: global search (q, matched via the findings_fts FTS5 projection)\ncombined with category/source filters. Omitting all three params returns\nevery finding in the job (CP4's original unfiltered behavior).",
        "operationId": "list_findings_api_jobs__job_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "collector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collector"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FindingRead"
                  },
                  "title": "Response List Findings Api Jobs  Job Id  Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Findings",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/subdomains": {
      "get": {
        "operationId": "list_subdomains_api_jobs__job_id__subdomains_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubdomainRowRead"
                  },
                  "title": "Response List Subdomains Api Jobs  Job Id  Subdomains Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subdomains",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/subdomains.csv": {
      "get": {
        "description": "FR-11: formula-safe UTF-8 CSV, generated entirely from persisted\nevidence - no provider calls.",
        "operationId": "export_subdomains_csv_api_jobs__job_id__subdomains_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Subdomains Csv",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}